requirements.txt
xlrd==1.2.0
azure-storage==0.36.0
azure-keyvault==1.1.0